Fix issues with the nav menu
Fix for a case that there is a secondary version of the menu item
 in the menu.( i.e someone makes two of the same menu item but one
 has two word string and the second has a one word string.)

Split up the text and loop through array of the string to look for
secondary version of a multi-word nav bar item.

Make sure that only the first item with the nav bar name gets the extra
class added. Since items in the sub-menu might have a similar name.